returned type of fibonacci method changed from int to long to cover more results

+++ src/main/java/pl/falcor/fibonacci/Fibonacci.java
@@ -3,13 +3,13 @@ package pl.falcor.fibonacci;
 class Fibonacci {
 
     private int inputValue;
-    private int result;
+    private long result;
 
     public int getInputValue() {
         return inputValue;
     }
 
-    public int getResult() {
+    public long getResult() {
         return result;
     }
 
@@ -21,7 +21,7 @@ class Fibonacci {
         this.result = fibonacci(inputValue);
     }
 
-    private int fibonacci(int n) {
+    private long fibonacci(int n) {
         if (n <= 1) return n;
         else return fibonacci(n - 2) + fibonacci(n - 1);
     }
